requirements.txt
Django==2.0.6
django-cors-headers==2.2.0
djangorestframework==3.10.3
Pillow==6.1.0
pytz==2018.4
